+++ packages/create-vue-ssg/templates/docs/env.d.ts
@@ -1,1 +1,7 @@
 /// <reference types="vite/client" />
+
+declare module 'virtual:docs-markdown-playground/*' {
+  import type { Component } from 'vue'
+  const component: Component
+  export default component
+}