+++ packages/create-vue-ssg/templates/docs/src/content/playgroundRegistry.generated.ts
@@ -1,0 +1,9 @@
+import type { Component } from 'vue'
+
+export const docsVirtualPlaygroundRegistry: Record<string, Component> = {}
+
+export const docsVirtualPlaygroundSourceRegistry: Record<string, string> = {}
+
+export const docsVirtualPlaygroundSourceFileRegistry: Record<string, string> = {}
+
+export const docsVirtualPlaygroundSourceLanguageRegistry: Record<string, string> = {}